+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+- `^` uses real path for non-integer and large exponents.
 - Error handling for builtin real functions.
 
 ### Fixed